06
Implement methods to render the individual numbers that may appear in a
cell. 1 should be blue, 2 should be green, 3 should be yellow, and four
and above should be red. Also write code to determine the number that
should be displayed in each cell when the game is initialised for the
first time